+++ search/domain/service.go
@@ -63,11 +63,9 @@ type (
 		// SelectedDesc true if sorting by this field is active
 		SelectedDesc bool
 		// Asc - represents the field that is used to trigger ascending search.
-		// Deprecated: use "Field" and "SelectedAsc" instead to set which field should be sortable
-		Asc string // Should it be deprecated ?? marked as deprecated in 2019 and used in 2020
+		Asc string
 		// Desc - represents the field that is used to trigger descending search.
-		// Deprecated: use "Field" and "SelectedDesc" instead to set which field should be sortable
-		Desc string // Should it be deprecated ?? marked as deprecated in 2019 and used in 2020
+		Desc string
 	}
 
 	// FacetType for type facets