+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## v3.12.0 [upcoming]
+**checkout**
+* add config `commerce.checkout.orderService.skipCartValidation` to disable cart validation that happens right before placing an order in the OrderService
 
 ## v3.11.0