+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 
-## v3.9.0 [upcoming]
+## v3.10.0 [upcoming]
+
+## v3.9.0
+
 **search**
 * Introduce an Action to search result to advise the frontend to perform an action for that search different from displaying the result page.
   The action model is completely open to anything, but the most common use case is to redirect certain searches directly to product detail pages or special brand landing pages.